fix(strace): fix strace problem which will break kernel heap allocation
This strace problem is very weird, it will happen for await syscalls
strace. And it will break heap allocation somehow after. This is so
fucking weird that I have ever seen. I can not understand what is wrong
with this strace! Finally, I have to move sepc out of strace to let the
bug not happen.