fix(smp): fix bug that current_task return a ref that points at wrong task when smp is running
current_task returns a ref, that points to the task of local hart.
However, when smp is running, the schedule may change to a different
hart, in this way, `let task = current_task;`, this `task` variable
still points to the origin hart's task, but origin hart and hart running
now may be different, which will cause trouble.