fix(do_exit): SIGCHLD should be signaled to parant only if this is the last thread
If SIGCHLD is sent because of leader called do_exit, then parent will
recycle the leader in sys_wait4 immediately, which will cause other
threads in the same thread group without leader thread, which will cause
panic.
Therefore SIGCHLD should only be sent only if this is the last thread of
the thread group.