fix(buffer_cache): block cache is enabled to transfer to file cache now
Some block caches are mis recognized, e.g. directory. When a directory
is frequently accessed, it is likely to be recognized as a block cache.
However, when the directory is removed, file may be created on the same
block, and the block is now as a file cache block now.